requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
simplejson==3.17.5
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
